requirements.txt
asgiref==3.5.0
Django==2.1.5
django-ckeditor==6.2.0
django-decouple==2.1
django-fontawesome-5==1.0.18
django-js-asset==1.2.2
fontawesome==5.10.1.post1
Khayyam==3.0.17
Pillow==9.0.1
pytz==2021.3
sqlparse==0.4.2
